move detector classes to control>vision
