+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/control/vision/AprilTagDetector.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.subsystems.utilities.sensors;
+package org.firstinspires.ftc.teamcode.control.vision;
 
 import static org.firstinspires.ftc.teamcode.opmodes.MainAuton.mTelemetry;
 
@@ -6,7 +6,6 @@ import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.teamcode.control.vision.AprilTagDetectionPipeline;
 import org.openftc.apriltag.AprilTagDetection;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
