+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/control/vision/TeamPropDetector.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.subsystems.utilities.sensors;
+package org.firstinspires.ftc.teamcode.control.vision;
 
 import static org.firstinspires.ftc.teamcode.control.vision.PropDetectPipeline.Randomization.RIGHT;
 import static org.firstinspires.ftc.teamcode.opmodes.MainAuton.mTelemetry;
@@ -9,7 +9,6 @@ import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.teamcode.control.vision.PropDetectPipeline;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
