+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/mechanismtests/TestTeamPropDetector.java
@@ -28,7 +28,7 @@ import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.subsystems.utilities.sensors.TeamPropDetector;
+import org.firstinspires.ftc.teamcode.control.vision.TeamPropDetector;
 
 @TeleOp(group = "Single mechanism test")
 public final class TestTeamPropDetector extends LinearOpMode {
